requirements.txt
python-fasthtml
markdown
uvicorn>=0.29
claudette